+++ AddOns/BtWQuestsMidnight/Database/Objects.lua
@@ -1,0 +1,26 @@
+----- AUTO GENERATED - DO NOT EDIT
+
+BtWQuestsDatabase:AddObjectsTable({
+    [504349] = {
+        name = "Cracked Holokey",
+        locations = {
+            [2405] = {
+                {
+                    x = 0.357249,
+                    y = 0.791898,
+                },
+            },
+        },
+    },
+    [573714] = {
+        name = "Ani's Trinket Bag",
+        locations = {
+            [2437] = {
+                {
+                    x = 0.487032,
+                    y = 0.660701,
+                },
+            },
+        },
+    },
+})