+++ AddOns/BtWQuestsMidnight/Localization.zhCN.lua
@@ -1,0 +1,9 @@
+----- AUTO GENERATED - DO NOT EDIT
+
+if GetLocale() ~= "zhCN" then
+    return
+end
+
+local L = BtWQuests.L
+L["THE_DARKENING_SKY"] = "黑暗的天空"
+L["THE_LIGHTS_SUMMONS"] = "圣光的召唤"