+++ AddOns/BtWQuestsMidnightPrologue/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v1.1
+
+- Updated for Midnight release
+
 # v1.0
 
 - Added the The Cult Within questline